+++ packages/plugin-constraints/sources/Constraints.ts
@@ -3,7 +3,7 @@
 import {Ident, MessageName, Project, ReportError, Workspace} from '@yarnpkg/core';
 import {miscUtils, structUtils}                              from '@yarnpkg/core';
 import {xfs, ppath, PortablePath}                            from '@yarnpkg/fslib';
-// @ts-ignore
+// @ts-expect-error
 import plLists                                               from 'tau-prolog/modules/lists';
 import pl                                                    from 'tau-prolog';