+++ CHANGELOG.md
@@ -16,6 +16,13 @@
 - The progress bars will be properly styled when using the new Windows terminal on certain days.
 - Yarn will now avoid using deprecated versions of the dependencies, unless only deprecated versions are available for the requested ranges.
 - Build keys are now properly computed, which fixes issues where build scripts weren't always triggered when they should have been.
+- Negated glob patterns in the `workspace` field will now be processed correctly.
+- Yarn will now allow relative paths inside the `workspace:` protocol to start with `./`
+- Yarn will now show the actual error when it fails to resolve a request during `yarn add` and `yarn up`
+- The portable shell will now support calling `cd` and `exit` without arguments
+- Yarn will now show the exit code when a lifecycle script fails
+- Yarn's portable shell will now also pipe the stderr when using the `|&` pipeline
+- Yarn's portable shell will now respect the left associativity of list operators
 
 ### CLI
 
@@ -31,6 +38,9 @@
 ### Miscellaneous
 
 - Scripts can now use the `$RANDOM` variable as well as simple calculations using `+`, `-`, `*`, `/` and `()` inside `$(())`
+- Scripts can now use grouping curly braces (`{echo foo}`) to execute a command in the context of the current shell (without creating a subshell like in the case of `(echo foo)`).
+- Scripts can now end with a semicolon.
+- PnP linker will not remove lingering node_modules inside folders matching `pnpIgnorePatterns`
 
 ### Third-party integrations